alex-companion-v2/requirements.txt
fastapi
uvicorn
python-multipart
google-genai
python-dotenv
edge-tts
aiofiles
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
duckduckgo-search
python-docx
pypdf
